API/requirements.txt
# Core Django
Django==5.1.8
asgiref==3.8.1
sqlparse==0.5.3

# REST Framework and Auth
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

# Database
psycopg2-binary==2.9.10

# Service Libraries
PSNAWP==2.2.0
gunicorn==23.0.0

# HTTP and API
requests==2.32.3
requests-ratelimiter==0.7.0
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
pycountry==24.6.1
pyrate-limiter==2.10.0
typing_extensions==4.12.2
attrs==24.2.0
cryptography==44.0.3